cavs: timer: use 64bit maths for checking MIN_DELAY
32 bit can overflow, meaning that next < curr can occur and cause
an erroneous calculation of "next".

+++ drivers/timer/cavs_timer.c
@@ -141,7 +141,7 @@ void z_clock_set_timeout(int32_t ticks, bool idle)
 	cyc = (cyc / CYC_PER_TICK) * CYC_PER_TICK;
 	next = last_count + cyc;
 
-	if (((uint32_t)next - (uint32_t)curr) < MIN_DELAY) {
+	if ((next - curr) < MIN_DELAY) {
 		next += CYC_PER_TICK;
 	}
